+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * `zone-id?` predicate
 * `set-clock!` - sets the mocked clock value to the supplied instant
+* `when-joda-time-loaded` - macro which runs code when Joda-Time is on the classpath
 
 ## 0.3.1